Revert missing sample file
