+++ synthesizer/process/src/stack/helpers/mod.rs
@@ -16,4 +16,5 @@ use super::*;
 
 mod initialize;
 mod matches;
+mod sample;
 mod synthesize;
